Fix gap_epoch wrap-around bug by upgrading to uint32_t

+++ src/core/order_state.hpp
@@ -75,7 +75,8 @@ struct OrderState {
     MismatchMask current_mismatch{};      // 1-byte mask
 
     std::uint32_t timer_generation{0};    // generation-based lazy cancel
-    std::uint16_t gap_suppression_epoch{0};  // Upgraded to uint16_t to avoid wrap-around issues
+    std::uint32_t gap_suppression_epoch{0};  // Using uint32_t to match SequenceTracker::gap_epoch
+                                             // and avoid wrap-around issues (would require 4B+ gaps)
 
     // ===== Divergence emission tracking (FX-7053) =====
     // These fields support idempotent divergence emission to avoid flooding
